requirements.txt
nltk==3.6.2
Django==3.2.3
pandas==1.2.4
djangorestframework==3.12.4
sklearn==0.0
dill==0.3.3

numpy==1.19.5
tensorflow==2.5.0